+++ derive/src/attribute/mod.rs
@@ -1,0 +1,5 @@
+pub mod field;
+pub mod item;
+pub mod repr;
+
+pub use repr::Repr;